- Moved Dockerfile lines around - Now using CMD rather than ENTRYPOINT - Nextflow config now contains profiles - Adds usage notes to README - Configuration exposed in start.sh

+++ start.sh
@@ -1,21 +1,29 @@
 #!/bin/bash
 
 NF_WORKFLOW=main.nf
-NF_REPORT='-with-trace results/trace.txt'
-NF_OPTS='-with-docker centos:7'
+NF_REPORT=${NF_REPORT_OPTIONS:-'-with-trace'}
+NF_OPTS=${NF_OPTIONS:-''}
 NF_INIT=${NF_INIT_SCRIPT:-$HOME/nextflow-init.sh}
+NF_PROFILE=${NF_PROFILE_NAME:-standard}
 
 echo "Looking for init script $NF_INIT"
 if [ -f $NF_INIT ]; then
     echo "Executing init script"
     source $NF_INIT
+else
+    echo "No init script found"
 fi
 
 echo "Running Nextflow"
-nextflow run $NF_WORKFLOW $NF_OPTS $NF_REPORT
-echo 'Nextflow completed'
+nextflow run $NF_WORKFLOW -profile $NF_PROFILE $NF_OPTS $NF_REPORT
+echo "Nextflow completed (ExitCode=$?)"
 
-echo "Moving results..."
-mv results/* .
-rm -rf results
-echo "Results moved"
+# Optionally move results
+if [ -d "results" ]; then
+    echo "Moving results..."
+    mv results/* .
+    rm -rf results
+    echo "Results moved"
+else
+    echo "No results directory"
+fi
